Dotnet core build and test for RDKit
