+++ Code/JavaWrappers/csharp_wrapper/RDKitDotNetCore.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RDKitDotNetCoreWrapper", "swig_csharp\RDKitDotNetCoreWrapper.csproj", "{B97DDA0A-D8BF-4653-926B-F13DFC27FA7C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RDKitCSharpTest", "RdkitTests\RDKitCSharpTest.csproj", "{65A680C7-921C-47CF-9737-BF8726396B92}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B97DDA0A-D8BF-4653-926B-F13DFC27FA7C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B97DDA0A-D8BF-4653-926B-F13DFC27FA7C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B97DDA0A-D8BF-4653-926B-F13DFC27FA7C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B97DDA0A-D8BF-4653-926B-F13DFC27FA7C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{65A680C7-921C-47CF-9737-BF8726396B92}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{65A680C7-921C-47CF-9737-BF8726396B92}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{65A680C7-921C-47CF-9737-BF8726396B92}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{65A680C7-921C-47CF-9737-BF8726396B92}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
