+++ Code/JavaWrappers/csharp_wrapper/RdkitTests/MolDrawHighlightTest.cs
