+++ Code/JavaWrappers/csharp_wrapper/swig_csharp/RDKitDotNetCoreWrapper.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <Target Name="CopyCustomContent" AfterTargets="AfterBuild">
+	  <Copy SourceFiles="../RDKFuncs.dll" DestinationFolder="$(OutDir)" />
+  </Target>
+</Project>
